WIP: Insert applications table to local PSQL successfully
TODO: need to add packages and dependencies tables
