+++ .gitignore
@@ -10,6 +10,7 @@ __pycache__/
 
 # Distribution / packaging
 .Python
+log/
 build/
 develop-eggs/
 dist/
